fix(#2806): R10 anchored pass full-token scan, line-scoped historical exemption, dirty-SA exit ordering
