+++ crates/cli-sub-agent/src/pipeline_tests_no_op_gate.rs
@@ -39,6 +39,7 @@ fn build_test_ctx<'a>(
         turn_count: 0,
         output_tokens: None,
         sa_mode,
+        original_exit_code: None,
     }
 }
