+++ crates/cli-sub-agent/src/pipeline_tests_no_op_gate_2181.rs
@@ -38,6 +38,7 @@ fn build_commit_no_op_ctx<'a>(
         turn_count: 0,
         output_tokens: None,
         sa_mode: true,
+        original_exit_code: None,
     }
 }
